chuyển cách Extensions của Objects chính thức thành thư mục riêng, thêm tính năng mới,
Tính năng mới:
1- Lấy tiếp ảnh từ Id và skipImages ?, takeImages (Ví dụ lấy đã có 8 ảnh và cần lấy ảnh thứ 9 và 10 thì truyền vào Id, skip = 8, take = 2)

2- GetSomeByFields (dành cho intro) có 3 phần (số lượng cần lấy, skip id đã nhận từ dữ liệu, và dict lưu các Field tùy biến tìm kiếm cho IntroBook)
