+++ BE/ApiTruyenLau/Services/BookServices.cs
@@ -1,4 +1,5 @@
 ﻿using ApiTruyenLau.Services.Interfaces;
+using ApiTruyenLau.Objects.Extensions.Items;
 using DataConnecion.MongoObjects;
 using Newtonsoft.Json;
 using Item = ApiTruyenLau.Objects.Generics.Items;
@@ -8,6 +9,8 @@ using ItemCvt = ApiTruyenLau.Objects.Converters.Items;
 using Newtonsoft.Json.Serialization;
 using ApiTruyenLau.Objects.Converters.Items;
 using ZstdSharp;
+using Microsoft.AspNetCore.Mvc;
+using System.Drawing;
 
 
 namespace ApiTruyenLau.Services
@@ -84,9 +87,126 @@ namespace ApiTruyenLau.Services
 			}
 			catch (Exception ex) { throw new Exception(ex.Message); }
 		}
+
+		/// <summary>
+		/// Mỗi một thông điệp gửi 1 yêu cầu (ví dụ: genre thì chỉ lấy 1 kiểu cố định, không lấy theo nhiều loại cùng lúc) 
+		/// </summary>
+		/// <param name="amountIntros"></param>
+		/// <param name="bookFields"></param>
+		/// <returns></returns>
+		/// <exception cref="Exception"></exception>
+		public async Task<List<ItemCvt.IntroBookPartCvt>> GetIntrosBySomething(int amountIntros, List<string> skipIds, Dictionary<string, string> bookFields)
+		{
+			try
+			{
+				// bookFields của Api trả về có thể sẽ liên quan tới vấn đề viết hoa viết thường
+				// Chuyển tất cả các key trong bookFields về lowercase và đem đối chiếu với Properties của Book ở dạng lower case 
+				var lowerCaseBookFields = bookFields.ToDictionary(entry => entry.Key.ToLower(), entry => entry.Value);
+				var bookProperties = typeof(Item.Book).GetProperties().ToDictionary(prop => prop.Name.ToLower(), prop => prop.Name);
+				var validBookFields = lowerCaseBookFields
+					.Where(entry => bookProperties.ContainsKey(entry.Key))
+					.ToDictionary(entry => bookProperties[entry.Key], entry => entry.Value);
+				var findedBookObjs = await _DB.GetMongoDBEntity(typeof(Item.Book)).FindObjects(validBookFields);
+
+				if (findedBookObjs != null && findedBookObjs.Count > 0)
+				{
+					var settings = new JsonSerializerSettings
+					{
+						MissingMemberHandling = MissingMemberHandling.Ignore,
+						SerializationBinder = new MySerializationBinderBook()
+					};
+					var findedBooks = findedBookObjs
+						.Select(obj => JsonConvert.DeserializeObject<Item.Book>(JsonConvert.SerializeObject(obj), settings))
+						.ToList();
+					var filteredBooks = findedBooks.Where(book => !skipIds.Contains(book!.Id)); // Bỏ qua các cuốn sách có Id nằm trong skips
+					var resultBooks = filteredBooks.Take(amountIntros).ToList(); // ít hơn thì lấy tất 
+					return resultBooks.Select(book => book?.ToIntroBookPartCvt()).ToList()!; // không còn sách thì trả về lỗi hết sách 
+				}
+				throw new Exception("Không có sách nào");
+			}
+			catch (Exception ex) { throw new Exception(ex.Message); }
+		}
 		#endregion Phần intro sách
 
+		#region Nội dung sách 
+		/// <summary>
+		/// Cái này sẽ lấy toàn bộ nội dung sách nên cần cân nhắc trước khi dùng 
+		/// Hợp lý hơn cho việc sử dụng với truyện chữ 
+		/// </summary>
+		/// <param name="bookId"></param>
+		/// <returns></returns>
+		/// <exception cref="NotImplementedException"></exception>
+		/// <exception cref="Exception"></exception>
+		public async Task<ItemCvt.BookCvt> GetBookById(string bookId)
+		{
+			try
+			{
+				var findedBookObjs = await _DB.GetMongoDBEntity(typeof(Item.Book)).FindObjects(new Dictionary<string, string>()
+				{
+					{ nameof(Item.Book.Id), bookId}
+				});
+				if (findedBookObjs != null && findedBookObjs.Count > 0)
+				{
+					var settings = new JsonSerializerSettings
+					{
+						MissingMemberHandling = MissingMemberHandling.Ignore,
+						SerializationBinder = new MySerializationBinderBook()
+					};
+					Item.Book? findedBook = findedBookObjs
+						.Select(obj => JsonConvert.DeserializeObject<Item.Book>(JsonConvert.SerializeObject(obj), settings))
+						.ToList().ElementAt(0);
+					if (findedBook != null)
+						Console.WriteLine(findedBook.CoverLink);
+					return findedBook?.ToBookCvt()!;
+				}
+				throw new NotImplementedException();
+			}
+			catch (Exception ex) { throw new Exception(ex.Message); }
+		}
+
+		/// <summary>
+		/// Lấy ảnh tiếp theo của nội dung sách 
+		/// Giả sử đọc tới ảnh thứ 5 và cần lấy ảnh 6 và 7 thì skipImages = 5, takeImages = 2
+		/// </summary>
+		/// <param name="bookId"></param>
+		/// <param name="skipImages"></param>
+		/// <param name="takeImages"></param>
+		/// <returns></returns>
+		/// <exception cref="NotImplementedException"></exception>
+		/// <exception cref="Exception"></exception>
+		public async Task<List<string>> GetNextImagesForContent(string bookId, int skipImages, int takeImages)
+		{
+			try
+			{
+				var findedBookObjs = await _DB.GetMongoDBEntity(typeof(Item.Book)).FindObjects(new Dictionary<string, string>()
+				{
+					{ nameof(Item.Book.Id), bookId}
+				});
+				if (findedBookObjs != null && findedBookObjs.Count > 0)
+				{
+					var settings = new JsonSerializerSettings
+					{
+						MissingMemberHandling = MissingMemberHandling.Ignore,
+						SerializationBinder = new MySerializationBinderBook()
+					};
+					Item.Book? findedBook = findedBookObjs
+						.Select(obj => JsonConvert.DeserializeObject<Item.Book>(JsonConvert.SerializeObject(obj), settings))
+						.ToList().ElementAt(0);
+					return findedBook?.GetImageAtElementsStringBase64Png(skipImages, takeImages)!;
+				}
+				throw new NotImplementedException("Hết ảnh rồi");
+			}
+			catch (Exception ex) { throw new Exception(ex.Message); }
+		}
+		#endregion Nội dung sách 
+
 		#region Phần tạo sách
+		/// <summary>
+		/// Tạo mới sách bằng Json 
+		/// </summary>
+		/// <param name="bookCreaterCvts"></param>
+		/// <returns></returns>
+		/// <exception cref="Exception"></exception>
 		public async Task<bool> CreateBooks(List<ItemCvt.BookCreaterCvt> bookCreaterCvts)
 		{
 			try
