+++ BE/ApiTruyenLau/Services/Interfaces/IBookServices.cs
@@ -9,7 +9,14 @@ namespace ApiTruyenLau.Services.Interfaces
 		public Task<ItemCvt.IntroBookPartCvt> GetIntroById(string bookId); 
 		// có thể lấy theo ngẫu nhiên hoặc theo tương tác của người dùng 
 		public Task<ItemCvt.IntroBookPartCvt> GetIntros(string userId);
+		public Task<List<ItemCvt.IntroBookPartCvt>> GetIntrosBySomething(int amountIntros, List<string> skipIds, Dictionary<string, string> bookFields);
 		#endregion Phần intro sách
+
+		#region Nội dung sách 
+		public Task<ItemCvt.BookCvt> GetBookById(string bookId);
+		public Task<List<string>> GetNextImagesForContent(string bookId, int skipImages, int takeImages);
+		#endregion Nội dung sách 
+
 		#region Phần tạo sách
 		public Task<bool> CreateBooks(List<ItemCvt.BookCreaterCvt> bookCreaterCvts);
 		#endregion Phần tạo sách
